fixing bugs:checking if sub download directory  exist before clearing it && fixing a search bug when no movies are found, it remove everything not just movie section.

+++ main.js
@@ -62,7 +62,8 @@ app.on("ready", () => createWindow());
 
 app.on("window-all-closed", () => {
   if (closeWindow) app.quit();
-  fs.readdirSync(subDirectory).forEach(file => {fs.unlinkSync(path.join(subDirectory,file))});
+  if(fs.existsSync(subDirectory))
+    fs.readdirSync(subDirectory).forEach(file => {fs.unlinkSync(path.join(subDirectory,file))});
 });
 
 // ======================= IPC HANDLERS =======================
